[bugfix] address freeze when exiting to system
